Added Real OpenSSL AES256/ChaCha20 crpyto
- Replace XOR-based AES256/ChaCha20 with OpenSSL EVP implementations
- Add cryptographic security using FIPS-validated algorithms
